♻️ Hide service implementations behind public factories
